+++ src/ska_mid_dish_manager/component_managers/ds_cm.py
@@ -33,6 +33,7 @@ class DSComponentManager(TangoDeviceComponentManager):
             "operatingMode",
             "powerState",
             "healthState",
+            "healthInfo",
             "pointingState",
             "indexerPosition",
             "desiredPointingAz",